Add monitored proxy objecs
We can now creates a monitored Proxy object that can come "live" as soon as the
referenced DBus name is available.

When the name is available (i.e. connected), the object will have the exact
same behavior as a normal Proxy object.

When the name is not available, the object will raise an error when trying to
access properties or call methods of the Proxy object with the exception of the
name property.