Fix callback function signature in connect_signal (BREAKING CHANGE)
Although the documentation has always been stating that the callback function
taken by Proxy:connect_signal would take the arguments sent by the signal, it
was taking an *array* containing the arguments instead.

This behavior has been around for quite a while, and some client libraries are
actually using this. However, I believe that the correct behavior is to take
the arguments passed by the signal one by one and have decided to introduce
this BREAKING CHANGE.