NixOS: 22.11 -> 24.05, fix and refactor tests
I had not updated this in a while and with 24.05 my super simple tests failed
when importing lgi with

     attempt to index local 'gobject' (a boolean value)

The problem is in lgi/ffi.lua, but the error is unclear because the code does
not use assert[1].

Anyway, the problem is that lgi cannot find the GObject typelib file... and the
reason is that at some point in NixOS that file has been moved from
gobject-introspection to glib. They are both used in a normal installation and
the GI_TYPELIB_PATH env var is generated correctly... but in the tests this was
not the case.

I have also removed the commented-out lua51Packages because it's been like this
for a while and I don't really use Lua 5.1

The integration tests for Lua 5.2, 5.3 and LuaJIT all work.

[1] https://github.com/lgi-devs/lgi/issues/71